Add files via upload
for those that Datamax and Datamax O'Neil label printers that just need a little program to test to see if the communication between the computer (Windows OS) and the printer works.  causes the printer to do a self test print and then prints the actual com port it communicated on.  If you want to make donations to my work, let me know.